pyproject.toml
[tool.poetry]
name = "product_analyser"
version = "0.1.0"
description = ""
authors = ["varun-s22 <93476421+varun-s22@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
matplotlib = "^3.5.1"
requests = "^2.27.1"
pandas = "^1.4.1"
secure-smtplib = "^0.1.1"
python-dotenv = "^0.19.2"
lxml = "^4.7.1"
Flask = "^2.0.3"
textblob = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
